+++ src/transports/ses.ts
@@ -7,7 +7,7 @@
  * @example
  * ```ts
  * import { SESTransport } from "sently/transports/ses";
- * import { createMailer } from "sently";
+ * import { createMailer } from "sently/mailer";
  *
  * const mailer = await createMailer({
  *   transport: new SESTransport({